+++ packages/callwave_flutter_method_channel/android/src/main/kotlin/com/callwave/flutter/methodchannel/CallwaveConstants.kt
@@ -29,6 +29,7 @@ object CallwaveConstants {
     const val EXTRA_EXTRA = "extra"
     const val EXTRA_LAUNCH_ACTION = "launchAction"
     const val EXTRA_ACCEPTANCE_STATE = "acceptanceState"
+    const val EXTRA_CONNECTED_AT_MS = "connectedAtMs"
     const val EXTRA_OUTCOME_REASON = "outcomeReason"
     const val EXTRA_INCOMING_ACCEPT_STRATEGY = "incomingAcceptStrategy"
     const val EXTRA_BACKGROUND_DISPATCHER_HANDLE = "backgroundDispatcherHandle"